ecommerce/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
pytz==2021.1
requests==2.25.1
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
